Enable daemons on remote systems
This picks up the hostname from the default-directory and uses it to
name the buffers, so you get a distinct pair of daemons buffers for
every system you visit.

This in combination with the effort to de-state means daemons now
works on remote systems (including switching back and forth
seamlessly, which was the problem before).