Add rudimentary support for systemd user services
Add a variable `daemons-systemd-is-user`, which toggles all of the `systemctl`
commands to add the `--user` flag. This makes the list and all of the commands
operate on user services rather than the system ones.

Also add a command `daemons-systemd-toggle-user`, which will toggle the variable
and refresh the list. It is bound to `u` in `daemons-mode-map` if the `:test`
for `daemons-systemd` is `t`.

Closes #3