Remove breaking niceness toggle
When defining the keymap, a check for if the submodule we're loading was
'systemd was added. This was intended as a niceness feature so that those that
don't run systemd wouldn't have something that won't do anything this key.

Unfortunately this broke the build because the submodules are not defined by the
time we're setting the keymap. To avoid this and to not complicate the code by
moving unrelated things around, we simply always add the key. Functionally, it
will only run (revert-buffer) in other modes, so it should not be a bother.